requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
WTForms==3.0.1 